MLB_pitchers_data
I downloaded  Statcast and Fangraphs baseball data specifically on pitchers from the past 11 years separately.  The above coding is what I have completed thus far.  The goal of the experiment is to conclude if there is a correlation between any natural ability of a pitcher and the ability to strike batters out at a high rate.